pyproject.toml
[project]
name = "5-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-debug-toolbar>=6.0.0",
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "django-requestlogs>=0.8.3",
    "django-cors-headers>=4.9.0",
]
